Rename schemas/chats -> chat
