+++ imports/lib/models/chats.ts
@@ -1,9 +1,9 @@
 import { huntsMatchingCurrentUser } from '../../model-helpers';
-import ChatMessagesSchema, { ChatMessageType } from '../schemas/chats';
+import ChatMessageSchema, { ChatMessageType } from '../schemas/chat';
 import Base from './base';
 
 const ChatMessages = new Base<ChatMessageType>('chatmessages');
-ChatMessages.attachSchema(ChatMessagesSchema);
+ChatMessages.attachSchema(ChatMessageSchema);
 ChatMessages.publish(huntsMatchingCurrentUser);
 
 export default ChatMessages;
